requirements.txt
accelerate>=1.2.1
diffusers>=0.32.0
matplotlib>=3.9.0
peft>=0.14.0
Pillow>=11.0.0
safetensors>=0.5.0
torch>=2.5.0
torchvision>=0.20.0
tqdm>=4.66.0
transformers>=4.47.0
